requirements.txt
alembic==1.0.10
certifi==2019.3.9
chardet==3.0.4
Click==7.0
docopt==0.6.2
Flask==1.0.2
flask-marshmallow==0.10.0
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.9
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.2
mysqlclient==1.4.2.post1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.24.3
Werkzeug==0.15.2
yarg==0.1.9
